requirements.txt
torch
torchvision
numpy
opencv-python
PyYaml
lmdb
imgaug
pyclipper
python-Levenshtein
Pillow